add actions to send to screen

+++ lib/util/format_utils.dart
@@ -5,7 +5,7 @@ String formatTokenNumber(String src) {
     return '';
   }
 
-  int tokenNumber = int.parse(src);
+  BigInt tokenNumber = BigInt.parse(src);
   double formattedNumber = tokenNumber.toDouble() / 1000000000.0;
   return formattedNumber.toString();
 }
